refactor: 이미지 다운로드 커스텀 훅으로 변경
